Ability to link a dashboard to a sitemap
This introduces a "link to sitemap" option for dashboards, accessible
in the dashboard tile settings. When linked to a sitemap, the dashboard
will act a standard sitemap explorer, with pages etc.

Widgets cannot be added or removed in this mode and the binding to
items cannot be changed, but they may be moved around and layout
options (colors, appearance settings...) may be changed.

CAUTION:
When reordering the sitemap's widgets or adding new frames/pages,
unexpected things might happen - widgetIds are attributed sequen-
tially and are not stable if the sitemap changes.

Other known bugs and limitations:
- Mappings for switches are not supported
- Selection, Webview, Mapview, Video are not supported
- Setpoints are rendered as knobs
- Conditional visibility or colors are not supported
- Formatting of labels are not supported

Signed-off-by: Yannick Schaus <habpanel@schaus.net>