Experimental feature: custom widgets on main menu
This provides the ability to use a custom widget as a main menu tile, including
setting its config, thus allowing dynamic content on the main menu.

The widgets can be clicked like a normal tile, so no need to handle click events
in the widget's template - in fact, any kind of interactivity is discouraged.

You can use `ngModel.name` in the widget template to retrieve the dashboard label.

Also added the ability to prevent the tile from "brightening up" when clicked.

Removed the $interval for the old clock in the main menu controller.

Closes #34, #108.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>