Support for international date/time and number formats
Resolves #131 (in part).

This adds locale support to HABPanel for more than 150 languages and regions.
The locale configured in ESH (org.eclipse.smarthome.core.i18nprovider) is used.

Not all contents of angular-locale are included to limit the size overhead:
only the region-less variant of each language is considered except for the
following locales (arbitrary list, more can be added later):

- es-ar
- de-at
- en-au
- fr-be
- es-bo
- pt-br
- en-ca
- fr-ca
- fr-ch
- es-co
- en-gb
- en-hk
- zh-hk
- en-ie
- en-in
- fr-lu
- es-mx
- en-nz
- en-sg
- zh-sg
- es-us
- zh-tw
- en-za

Note: some dependencies in vendor.js were updated in the process.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>