Add angular-translate infrastucture for internationalization
Add translation support to almost every string in the app.

Separate into 3 parts: main, admin, widgets.

Add a custom service & filter to leave the English text as-is
but still allow a translation.

Partial translations are supported as well: if a key is not
present in the translations, the English text will be displayed
instead.

Complete French translation.
German translation skeleton.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>