Fix NaN display when dimmer issued ON/OFF command
The value of a slider/dimmer item is naively assumed to be numerical. Many OpenHAB dimmer components are issued ON/OFF commands instead of 0/100. This causes the slider widget to display NaN as the item value, and causes visual artifacts as well (value below the floor).

Add a check for ON/OFF strings, before assuming value is numerical and parsing it as a float.

Signed-off-by: Mike Nemat <michael.nemat@gmail.com> (github: mikenemat)